+++ src/TrivyOperator.Dashboard/Application/Trivy/Models/RbacAssessmentReportDto.cs
@@ -8,10 +8,10 @@ public class RbacAssessmentReportDto
     public Guid Uid { get; init; } = Guid.NewGuid();
     public string ResourceName { get; init; } = string.Empty;
     public string ResourceNamespace { get; init; } = string.Empty;
-    public long CriticalCount { get; init; }
-    public long HighCount { get; init; }
-    public long MediumCount { get; init; }
-    public long LowCount { get; init; }
+    public int CriticalCount { get; init; }
+    public int HighCount { get; init; }
+    public int MediumCount { get; init; }
+    public int LowCount { get; init; }
     public DateTime? CreationTimestamp { get; init; }
     public RbacAssessmentReportDetailDto[] Details { get; set; } = [];
 }
